==> Jupyter notebooks/4.Confirm INCEpTION positive labels.ipynb <==
# 4. Confirm INCEpTION positive labels

Manual quality check of annotations

import pandas as pd

folds_val = pd.read_csv(r'.\Intermediate results\folds_val.csv', sep = '\t')
folds_test = pd.read_csv(r'.\Intermediate results\folds_test.csv', sep = '\t')

#Validation

new_folds=folds_val

sentence_validations = []

for index, row in new_folds.iterrows():
    if row['relevance_manual'] <2:#== 1:
        while True:
            try:
                print("Index:", index)
                print('Sentence:', row['sentence'], 'Label:', row['relevance_manual'])
                user_input = int(input("Please enter an integer: "))
                break
            except:
                print("TRY AGAIN:")
    elif row['relevance_manual'] == 0:
        print('Sentence:', row['sentence'], 'Label:', row['relevance_manual'])
        user_input = 0
    sentence_validations.append(user_input)
    
new_folds['validated'] = sentence_validations 


new_folds.to_csv(r'.\Intermediate results\val_folds_validated.csv', sep = '\t')

#Test

new_folds=folds_test

sentence_validations = []

for index, row in new_folds.iterrows():
    if row['relevance_manual'] <2:#== 1:
        while True:
            try:
                print("Index:", index)
                print('Sentence:', row['sentence'], 'Label:', row['relevance_manual'])
                user_input = int(input("Please enter an integer: "))
                break
            except:
                print("TRY AGAIN:")
    elif row['relevance_manual'] == 0:
        print('Sentence:', row['sentence'], 'Label:', row['relevance_manual'])
        user_input = 0
    sentence_validations.append(user_input)
    
new_folds['validated'] = sentence_validations 


new_folds.to_csv(r'.\Intermediate results\test_folds_validated.csv', sep = '\t')
